WIP migration to local server client to allow for easier gui
